outstanding integration tests + refactoring
