+++ Makefile
@@ -1,0 +1,2 @@
+test:
+	mvn verify jacoco:report
