+++ src/main/java/uk/gov/ons/census/caseapisvc/config/AppConfig.java
@@ -1,0 +1,17 @@
+package uk.gov.ons.census.caseapisvc.config;
+
+import ma.glasnost.orika.MapperFacade;
+import ma.glasnost.orika.MapperFactory;
+import ma.glasnost.orika.impl.DefaultMapperFactory;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class AppConfig {
+  @Bean
+  public MapperFacade mapperFacade() {
+    MapperFactory mapperFactory = new DefaultMapperFactory.Builder().build();
+
+    return mapperFactory.getMapperFacade();
+  }
+}
