+++ src/main/java/uk/gov/ons/census/caseapisvc/model/dto/EmptyResponseDTO.java
@@ -1,0 +1,6 @@
+package uk.gov.ons.census.caseapisvc.model.dto;
+
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+
+@JsonSerialize
+public class EmptyResponseDTO {}
