+++ src/main/java/uk/gov/ons/census/caseapisvc/model/dto/EventDTO.java
@@ -1,0 +1,20 @@
+package uk.gov.ons.census.caseapisvc.model.dto;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import java.util.Date;
+import java.util.UUID;
+import lombok.Data;
+
+@Data
+public class EventDTO {
+
+  private UUID id;
+
+  private String category;
+
+  @JsonProperty("createdDateTime")
+  private Date eventDate;
+
+  @JsonProperty("description")
+  private String eventDescription;
+}
