+++ src/main/java/uk/gov/ons/census/caseapisvc/model/dto/ResponseDTO.java
@@ -1,0 +1,9 @@
+package uk.gov.ons.census.caseapisvc.model.dto;
+
+import lombok.Data;
+
+@Data
+public class ResponseDTO {
+  private String dateTime;
+  private String inboundChannel;
+}
