+++ src/main/java/uk/gov/ons/census/caseapisvc/model/entity/CaseState.java
@@ -1,0 +1,5 @@
+package uk.gov.ons.census.caseapisvc.model.entity;
+
+public enum CaseState {
+  ACTIONABLE
+}
