--- src/main/java/uk/gov/ons/census/casesvc/Application.java
@@ -1,11 +1,0 @@
-package uk.gov.ons.census.casesvc;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Application {
-  public static void main(String[] args) {
-    SpringApplication.run(Application.class);
-  }
-}
