--- src/main/java/uk/gov/ons/census/casesvc/config/AppConfig.java
@@ -1,28 +1,0 @@
-package uk.gov.ons.census.casesvc.config;
-
-import ma.glasnost.orika.MapperFacade;
-import ma.glasnost.orika.MapperFactory;
-import ma.glasnost.orika.impl.DefaultMapperFactory;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import uk.gov.ons.ctp.common.error.RestExceptionHandler;
-
-@Configuration
-public class AppConfig {
-  /**
-   * Rest Exception Handler
-   *
-   * @return a Rest Exception Handler
-   */
-  @Bean
-  public RestExceptionHandler restExceptionHandler() {
-    return new RestExceptionHandler();
-  }
-
-  @Bean
-  public MapperFacade mapperFacade() {
-    MapperFactory mapperFactory = new DefaultMapperFactory.Builder().build();
-
-    return mapperFactory.getMapperFacade();
-  }
-}
