--- src/main/java/uk/gov/ons/census/casesvc/config/WebSecurityConfig.java
@@ -1,25 +1,0 @@
-package uk.gov.ons.census.casesvc.config;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-
-@Configuration
-@EnableWebSecurity
-@Profile("test")
-public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
-
-  @Autowired
-  protected void configure(AuthenticationManagerBuilder auth) {
-    // ...
-  }
-
-  @Override
-  protected void configure(HttpSecurity httpSecurity) throws Exception {
-    httpSecurity.authorizeRequests().antMatchers("/").permitAll();
-  }
-}
