--- src/main/java/uk/gov/ons/census/casesvc/error/ApiError.java
@@ -1,16 +1,0 @@
-package uk.gov.ons.census.casesvc.error;
-
-import lombok.Data;
-import org.springframework.http.HttpStatus;
-
-@Data
-public class ApiError {
-  private HttpStatus status;
-
-  private String message;
-
-  public ApiError(HttpStatus status, String message) {
-    this.status = status;
-    this.message = message;
-  }
-}
