--- src/main/java/uk/gov/ons/census/casesvc/model/dto/EmptyResponseDTO.java
@@ -1,6 +1,0 @@
-package uk.gov.ons.census.casesvc.model.dto;
-
-import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-
-@JsonSerialize
-public class EmptyResponseDTO {}
