--- src/main/java/uk/gov/ons/census/casesvc/model/dto/EventDTO.java
@@ -1,20 +1,0 @@
-package uk.gov.ons.census.casesvc.model.dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import java.util.Date;
-import java.util.UUID;
-import lombok.Data;
-
-@Data
-public class EventDTO {
-
-  private UUID id;
-
-  private String category;
-
-  @JsonProperty("createdDateTime")
-  private Date eventDate;
-
-  @JsonProperty("description")
-  private String eventDescription;
-}
