--- src/main/java/uk/gov/ons/census/casesvc/model/dto/ResponseDTO.java
@@ -1,9 +1,0 @@
-package uk.gov.ons.census.casesvc.model.dto;
-
-import lombok.Data;
-
-@Data
-public class ResponseDTO {
-  private String dateTime;
-  private String inboundChannel;
-}
