--- src/main/java/uk/gov/ons/census/casesvc/model/entity/CaseState.java
@@ -1,5 +1,0 @@
-package uk.gov.ons.census.casesvc.model.entity;
-
-public enum CaseState {
-  ACTIONABLE
-}
