--- src/main/java/uk/gov/ons/census/casesvc/model/entity/Event.java
@@ -1,21 +1,0 @@
-package uk.gov.ons.census.casesvc.model.entity;
-
-import java.util.Date;
-import java.util.UUID;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.ManyToOne;
-import lombok.Data;
-
-@Data
-@Entity
-public class Event {
-  @Id private UUID id;
-
-  @ManyToOne private UacQidLink uacQidLink;
-
-  @Column private Date eventDate;
-
-  @Column private String eventDescription;
-}
