--- src/main/resources/import.sql
@@ -1,17 +1,0 @@
-INSERT INTO casev2.cases (case_ref, abp_code, action_plan_id, address_level, address_line1, address_line2, address_line3, address_type, arid, case_id, collection_exercise_id, estab_arid, estab_type, htc_digital, htc_willingness, lad, latitude, longitude, lsoa, msoa, oa, organisation_name, postcode, rgn, state, town_name, treatment_code, uprn) VALUES (10000000, 'RD06', 'e1f27112-107c-41e9-99a7-01abf343f76c', 'U', 'Flat 56 Francombe House', 'Commercial Road', null, 'HH', 'DDR190314000000195675', 'c0d4f87d-9d19-4393-80c9-9eb94f69c460', '045bbaf5-f219-4306-9593-0a7520eea2a8', 'DDR190314000000113740', 'Household', '5', '1', 'E06000023', '51.4463421', '-2.5924477', 'E01014540', 'E02003043', 'E00073438', null, 'XX1 0XX', 'E12000009', 'ACTIONABLE', 'Windleybury', 'HH_LF3R2E', null);
-INSERT INTO casev2.cases (case_ref, abp_code, action_plan_id, address_level, address_line1, address_line2, address_line3, address_type, arid, case_id, collection_exercise_id, estab_arid, estab_type, htc_digital, htc_willingness, lad, latitude, longitude, lsoa, msoa, oa, organisation_name, postcode, rgn, state, town_name, treatment_code, uprn) VALUES (10000001, 'RD06', 'e1f27112-107c-41e9-99a7-01abf343f76c', 'U', 'First And Second Floor Flat', '39 Cranbrook Road', null, 'HH', 'DDR190314000000239595', '16d79007-9224-448a-9e59-944d9d153fa1', '045bbaf5-f219-4306-9593-0a7520eea2a8', 'DDR190314000000060908', 'Household', '4', '2', 'E06000023', '51.4721166', '-2.5970579', 'E01014669', 'E02003031', 'E00074083', null, 'XX1 0XX', 'E12000009', 'ACTIONABLE', 'Windleybury', 'HH_LF3R3AE', null);
-INSERT INTO casev2.cases (case_ref, abp_code, action_plan_id, address_level, address_line1, address_line2, address_line3, address_type, arid, case_id, collection_exercise_id, estab_arid, estab_type, htc_digital, htc_willingness, lad, latitude, longitude, lsoa, msoa, oa, organisation_name, postcode, rgn, state, town_name, treatment_code, uprn) VALUES (10000002, 'RD06', 'e1f27112-107c-41e9-99a7-01abf343f76c', 'U', 'Flat 2', '60 Coronation Road', null, 'HH', 'DDR190314000000320079', '7f043f01-2f84-48ee-a4a0-850735b09692', '045bbaf5-f219-4306-9593-0a7520eea2a8', 'DDR190314000000114781', 'Household', '3', '3', 'E06000023', '51.4451742', '-2.5975551', 'E01033361', 'E02003050', 'E00074236', null, 'XX1 0XX', 'E12000009', 'ACTIONABLE', 'Windleybury', 'HH_LF3R3BE', '123456789012345');
-INSERT INTO casev2.cases (case_ref, abp_code, action_plan_id, address_level, address_line1, address_line2, address_line3, address_type, arid, case_id, collection_exercise_id, estab_arid, estab_type, htc_digital, htc_willingness, lad, latitude, longitude, lsoa, msoa, oa, organisation_name, postcode, rgn, state, town_name, treatment_code, uprn) VALUES (10000004, 'RD06', 'e1f27112-107c-41e9-99a7-01abf343f76c', 'U', '18 Marshfield Park Cleeve Wood Road', null, null, 'HH', 'DDR190314000000522356', '26712d73-89a5-46bc-8626-4dc3f0ed906d', '045bbaf5-f219-4306-9593-0a7520eea2a8', 'DDR190314000000638000', 'Household', '1', '5', 'E06000025', '51.4958465', '-2.513198', 'E01014897', 'E02003111', 'E00075275', null, 'XX11 0XX', 'E12000009', 'ACTIONABLE', 'Windleybury', 'HH_LF2R3BE', '123456789012345');
-INSERT INTO casev2.cases (case_ref, abp_code, action_plan_id, address_level, address_line1, address_line2, address_line3, address_type, arid, case_id, collection_exercise_id, estab_arid, estab_type, htc_digital, htc_willingness, lad, latitude, longitude, lsoa, msoa, oa, organisation_name, postcode, rgn, state, town_name, treatment_code, uprn) VALUES (10000003, 'RD04', 'e1f27112-107c-41e9-99a7-01abf343f76c', 'U', '14 Blackhorse Road', null, null, 'HH', 'DDR190314000000516472', 'ef3e36fc-ca58-4d6a-b2cd-fb406d7c4b00', '045bbaf5-f219-4306-9593-0a7520eea2a8', 'DDR190314000000516472', 'Household', '2', '4', 'E06000025', '51.4603426', '-2.5122237', 'E01015005', 'E02003117', 'E00173063', null, 'XX11 0XX', 'E12000009', 'ACTIONABLE', 'Windleybury', 'HH_LFNR1E', '123456789012345');
-
-INSERT INTO casev2.uac_qid_link (id, qid, uac, unique_number, caze_case_ref) VALUES ('3ee075b6-5e94-4c91-9b34-a64f22228b91', '0120000000000100', 'qhb6c2vx7mdf6m9l', 1, 10000000);
-INSERT INTO casev2.uac_qid_link (id, qid, uac, unique_number, caze_case_ref) VALUES ('48bcbef4-3f9e-42fb-b403-63766d47189f', '0120000000000200', 'ngcpbr3qxlsqxh4c', 2, 10000001);
-INSERT INTO casev2.uac_qid_link (id, qid, uac, unique_number, caze_case_ref) VALUES ('82353afb-f3af-49af-80a9-2a2481123e2c', '0120000000000300', 'tmlhhfhlsh86s2mj', 3, 10000002);
-INSERT INTO casev2.uac_qid_link (id, qid, uac, unique_number, caze_case_ref) VALUES ('1c44db04-fd7e-4d09-82c8-322a4fe622e3', '0120000000000400', 'vhf5l6v87ghm6r8d', 4, 10000003);
-INSERT INTO casev2.uac_qid_link (id, qid, uac, unique_number, caze_case_ref) VALUES ('9ad64c52-ffe4-4fba-b7d7-866760788ea4', '0120000000000500', 'tchsmkt5tj733kbk', 5, 10000004);
-
-INSERT INTO casev2.event (id, event_date, event_description, uac_qid_link_id) VALUES ('401ac27a-5896-462e-a935-54d871c13b17', '2019-05-01 13:17:13.033000', 'Case created', '3ee075b6-5e94-4c91-9b34-a64f22228b91');
-INSERT INTO casev2.event (id, event_date, event_description, uac_qid_link_id) VALUES ('cc23f012-2bfa-4067-9aef-9e97d9979882', '2019-05-01 13:17:13.089000', 'Case created', '48bcbef4-3f9e-42fb-b403-63766d47189f');
-INSERT INTO casev2.event (id, event_date, event_description, uac_qid_link_id) VALUES ('51f6a6ef-35dc-443d-99b5-f42be7503821', '2019-05-01 13:17:13.105000', 'Case created', '82353afb-f3af-49af-80a9-2a2481123e2c');
-INSERT INTO casev2.event (id, event_date, event_description, uac_qid_link_id) VALUES ('a8c1104e-d261-4407-a8c3-7cb6999186f8', '2019-05-01 13:17:13.126000', 'Case created', '1c44db04-fd7e-4d09-82c8-322a4fe622e3');
-INSERT INTO casev2.event (id, event_date, event_description, uac_qid_link_id) VALUES ('7e69e016-2208-450d-ac5b-eb96701920b9', '2019-05-01 13:17:13.153000', 'Case created', '9ad64c52-ffe4-4fba-b7d7-866760788ea4');
